backend/requirements.txt
fastapi
uvicorn
motor
pydantic
python-dotenv
passlib[bcrypt]
bcrypt==4.0.1
email-validator
python-jose[cryptography]
python-multipart
requests